requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
python-decouple==3.8
psycopg2-binary==2.9.11
openpyxl==3.1.5
requests==2.32.5
APScheduler==3.11.2
Pillow==12.1.1
